+++ CHANGELOG.rst
@@ -24,8 +24,11 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
+- Fixed import when running in optimized mode (``PYTHONOPTIMIZE=2``).
+  `#373 <https://github.com/hynek/structlog/pull/373>`_
 - Added ``structlog.threadlocal.bound_threadlocal`` as a thread-local equivalent of ``structlog.threadlocal.tmp_bind``
   for temporarily binding keys to the global context through a decorator/context manager.
+ `#371 <https://github.com/hynek/structlog/pull/371>`_
 
 
 ----